refactor: 도메인 패키지 내부 클래스 이동
