+++ src/main/java/JanggiController.java
@@ -1,6 +1,7 @@
-import domain.*;
 import domain.board.Board;
+import domain.piece.Camp;
 import domain.piece.Piece;
+import domain.position.ElephantFormation;
 import domain.position.Position;
 import repository.JanggiRepository;
 import view.InputView;
