+++ src/main/java/domain/piece/Camp.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece;
 
 public enum Camp {
     HAN("한나라"),
