+++ src/main/java/domain/piece/Elephant.java
@@ -1,6 +1,5 @@
 package domain.piece;
 
-import domain.*;
 import domain.board.BoardReader;
 import domain.position.Direction;
 import domain.position.MoveDirection;
