+++ src/main/java/domain/piece/General.java
@@ -1,8 +1,5 @@
 package domain.piece;
 
-import domain.Camp;
-import domain.PieceType;
-
 public class General extends SingleStepPiece {
 
     public General(Camp camp) {
