+++ src/main/java/domain/piece/Guard.java
@@ -1,8 +1,5 @@
 package domain.piece;
 
-import domain.Camp;
-import domain.PieceType;
-
 public class Guard extends SingleStepPiece {
 
     public Guard(Camp camp) {
