+++ src/test/java/domain/piece/SoldierTest.java
@@ -1,7 +1,6 @@
 package domain.piece;
 
 import domain.board.Board;
-import domain.Camp;
 import domain.position.Position;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
